updates on model,utils,repository and controller

+++ annotationplus/src/main/java/com/nbu/annotationplus/repository/CategoryRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface CategoryRepository extends JpaRepository<Category, Long>  {
-    List<Category> findAllByUserId(int userId);
+    List<Category> findAllByUserId(Long userId);
 }
